@@ -6,8 +6,8 @@ authors = [
 ]
 description = "Image processing library for astrophotography"
 readme = "README.md"
-license = { file = "LICENSE": }
-requires-python = ">=3.7"
+license = { file = "LICENSE" }
+requires-python = ">=3.8"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: Apache License 2.0",
@@ -26,7 +26,7 @@ dependencies = [
 Homepage = "https://github.com/austinlucaslake/astrosight"
 
 [project.optional-dependencies]
-dev = ["black", "flake8", "pytest", "pip-tools"]
+dev = ["black", "autopep8", "flake8", "flake8-docstrings", "autoflake", "pytest", "pipreqs", "pip-tools"]
 
 [project.scripts]
 astrosight-cli = "astrosight.cli:main"
